backend/requirements.txt
# FastAPI and web server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Core Dependencies
python-dotenv>=1.0.0

# LangGraph for orchestration
langgraph>=0.2.0
langchain-core>=0.3.0
langchain>=0.3.0            # Required for LangGraph compatibility
langgraph-checkpoint-sqlite # For LangGraph state persistence

# LLM APIs
anthropic>=0.39.0           # Claude 4.5 Haiku for URL discovery
google-genai>=1.0.0         # Gemini 2.5 Flash for product confirmation

# Price scraping
apify-client>=1.0.0         # Apify API client for price scraping

# HTTP clients
requests>=2.31.0
httpx>=0.27.0               # Required by google-genai and anthropic SDKs

# Data validation and parsing
pydantic>=2.0.0

# NEW: Authentication packages
PyJWT==2.8.0
python-jose[cryptography]==3.3.0

# NEW: Database packages
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
alembic==1.13.0

# NEW: SMS/OTP service (for production)
twilio==8.11.0